+++ test/sparse.jl
@@ -1,0 +1,4 @@
+include("sparsedir/sparse.jl")
+include("sparsedir/umfpack.jl")
+include("sparsedir/cholmod.jl")
+include("sparsedir/spqr.jl")